+++ plugin/src/main/kotlin/org/jlleitschuh/gradle/ktlint/TaskCreation.kt
@@ -1,9 +1,12 @@
 package org.jlleitschuh.gradle.ktlint
 
+import org.gradle.api.Project
 import org.gradle.api.file.FileTree
+import org.gradle.api.tasks.TaskCollection
 import org.gradle.api.tasks.TaskProvider
 import org.gradle.language.base.plugins.LifecycleBasePlugin
 import org.jlleitschuh.gradle.ktlint.tasks.BaseKtLintCheckTask
+import org.jlleitschuh.gradle.ktlint.tasks.GenerateBaselineTask
 import org.jlleitschuh.gradle.ktlint.tasks.GenerateReportsTask
 import org.jlleitschuh.gradle.ktlint.tasks.KtLintCheckTask
 import org.jlleitschuh.gradle.ktlint.tasks.KtLintFormatTask
@@ -30,6 +33,18 @@ internal fun createFormatTask(
     .registerTask(
         KtLintFormatTask.buildTaskNameForSourceSet(sourceSetName)
     ) {
+        mustRunAfter(project.tasks.named(KtLintFormatTask.KOTLIN_SCRIPT_TASK_NAME))
+
+        val rootProjectName = project.rootProject.name
+        var parentProject: Project? = project.parent
+        while (parentProject != null && parentProject.name != rootProjectName) {
+            val parentProjectPath = parentProject.path
+            parentProject.plugins.withId("org.jlleitschuh.gradle.ktlint") {
+                mustRunAfter("$parentProjectPath:${KtLintFormatTask.KOTLIN_SCRIPT_TASK_NAME}")
+            }
+            parentProject = parentProject.parent
+        }
+
         description = KtLintFormatTask.buildDescription(".kt")
         configureBaseCheckTask(pluginHolder) {
             setSource(kotlinSourceDirectories)
@@ -125,6 +140,7 @@ internal fun <T : BaseKtLintCheckTask> createGenerateReportsTask(
 ): TaskProvider<GenerateReportsTask> = pluginHolder.target.registerTask(
     GenerateReportsTask.generateNameForSourceSets(sourceSetName, lintType)
 ) {
+    mustRunAfter(project.tasks.named(KtLintFormatTask.KOTLIN_SCRIPT_TASK_NAME))
     reportsName.set(GenerateReportsTask.generateNameForSourceSets(sourceSetName, lintType))
     commonConfiguration(pluginHolder, lintTask)
 }
@@ -159,4 +175,33 @@ private fun <T : BaseKtLintCheckTask> GenerateReportsTask.commonConfiguration(
     ignoreFailures.set(pluginHolder.extension.ignoreFailures)
     verbose.set(pluginHolder.extension.verbose)
     ktLintVersion.set(pluginHolder.extension.version)
+    @Suppress("UnstableApiUsage")
+    baseline.set(
+        pluginHolder.extension.baseline
+            .flatMap {
+                if (it.asFile.exists()) {
+                    pluginHolder.target.objects.fileProperty().apply { set(it.asFile) }
+                } else {
+                    pluginHolder.target.objects.fileProperty()
+                }
+            }
+    )
+}
+
+internal fun createGenerateBaselineTask(
+    pluginHolder: KtlintPlugin.PluginHolder,
+    lintTasks: TaskCollection<out BaseKtLintCheckTask>
+): TaskProvider<GenerateBaselineTask> = pluginHolder.target.registerTask(
+    GenerateBaselineTask.NAME
+) {
+    description = GenerateBaselineTask.DESCRIPTION
+    group = HELP_GROUP
+
+    dependsOn(lintTasks)
+
+    ktLintClasspath.setFrom(pluginHolder.ktlintConfiguration)
+    baselineReporterClasspath.setFrom(pluginHolder.ktlintBaselineReporterConfiguration)
+    discoveredErrors.from(lintTasks.map { it.discoveredErrors })
+    ktLintVersion.set(pluginHolder.extension.version)
+    baselineFile.set(pluginHolder.extension.baseline)
 }