+++ plugin/src/main/kotlin/org/jlleitschuh/gradle/ktlint/worker/KtLintWorkAction.kt
@@ -2,11 +2,14 @@ package org.jlleitschuh.gradle.ktlint.worker
 
 import com.pinterest.ktlint.core.KtLint
 import com.pinterest.ktlint.core.LintError
+import com.pinterest.ktlint.core.ParseException
 import com.pinterest.ktlint.core.RuleSet
 import com.pinterest.ktlint.core.RuleSetProvider
 import net.swiftzer.semver.SemVer
+import org.gradle.api.GradleException
 import org.gradle.api.file.ConfigurableFileCollection
 import org.gradle.api.file.RegularFileProperty
+import org.gradle.api.logging.Logging
 import org.gradle.api.provider.Property
 import org.gradle.api.provider.SetProperty
 import org.gradle.workers.WorkAction
@@ -15,6 +18,9 @@ import java.util.ServiceLoader
 
 @Suppress("UnstableApiUsage")
 abstract class KtLintWorkAction : WorkAction<KtLintWorkAction.KtLintWorkParameters> {
+
+    private val logger = Logging.getLogger("ktlint-worker")
+
     override fun execute() {
         val ruleSets = loadRuleSetsAndFilterThem(
             parameters.enableExperimental.getOrElse(false),
@@ -29,6 +35,8 @@ abstract class KtLintWorkAction : WorkAction<KtLintWorkAction.KtLintWorkParamete
         val userData = generateUserData()
         val debug = parameters.debug.get()
         val formatSource = parameters.formatSource.getOrElse(false)
+
+        resetEditorconfigCache()
 
         val result = mutableListOf<LintErrorResult>()
 
@@ -47,16 +55,24 @@ abstract class KtLintWorkAction : WorkAction<KtLintWorkAction.KtLintWorkParamete
                 }
             )
 
-            if (formatSource) {
-                val currentFileContent = it.readText()
-                val updatedFileContent = KtLint.format(ktLintParameters)
+            try {
+                if (formatSource) {
+                    val currentFileContent = it.readText()
+                    val updatedFileContent = KtLint.format(ktLintParameters)
 
-                if (updatedFileContent != currentFileContent) {
-                    it.writeText(updatedFileContent)
+                    if (updatedFileContent != currentFileContent) {
+                        it.writeText(updatedFileContent)
+                    }
+                } else {
+                    KtLint.lint(ktLintParameters)
                 }
-            } else {
-                KtLint.lint(ktLintParameters)
+            } catch (e: ParseException) {
+                throw GradleException(
+                    "KtLint failed to parse file: ${it.absolutePath}",
+                    e
+                )
             }
+
             result.add(
                 LintErrorResult(
                     lintedFile = it,
@@ -73,6 +89,14 @@ abstract class KtLintWorkAction : WorkAction<KtLintWorkAction.KtLintWorkParamete
                 result,
                 parameters.discoveredErrorsFile.asFile.get()
             )
+    }
+
+    private fun resetEditorconfigCache() {
+        if (parameters.editorconfigFilesWereChanged.get()) {
+            logger.info("Resetting KtLint caches")
+            // Calling trimMemory() will also reset internal loaded `.editorconfig` cache
+            KtLint.trimMemory()
+        }
     }
 
     private fun generateUserData(): Map<String, String> {
@@ -116,5 +140,6 @@ abstract class KtLintWorkAction : WorkAction<KtLintWorkAction.KtLintWorkParamete
         val formatSource: Property<Boolean>
         val discoveredErrorsFile: RegularFileProperty
         val ktLintVersion: Property<String>
+        val editorconfigFilesWereChanged: Property<Boolean>
     }
 }